requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
alembic==1.14.1
anyio==4.12.1
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
fastapi==0.128.0
feedparser==6.0.12
h11==0.16.0
idna==3.11
numpy==1.26.4
openai==1.60.0
opencv-python-headless==4.9.0.80
passlib[bcrypt]==1.7.4
piexif>=1.1.3
Pillow>=10.0.0
psycopg2-binary
pydantic==2.12.5
pydantic_core==2.41.5
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv
python-jose==3.3.0
python-multipart
python-whois==0.9.6
pytesseract
redis>=5.2.0
requests==2.32.5
resend
scipy==1.11.4
sentry-sdk[fastapi]==2.20.0
six==1.17.0
soundfile==0.12.1
sqlalchemy
starlette==0.50.0
supabase>=2.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
google-auth
boto3
firebase-admin
email-validator==2.1.1
librosa==0.10.2
